requirements.txt
flask
mysql-connector-python
openai
Werkzeug
python-dotenv
tika
argon2-cffi
PyJWT
tk
requests